Code cleanup.

+++ app/src/main/java/qa/dcsdr/diplomaticclub/Tools/Ellipsizer.java
@@ -8,7 +8,7 @@ public class Ellipsizer {
     private final static String NON_THIN = "[^iIl1\\.,']";
 
     private static int textWidth(String str) {
-        return (int) (str.length() - str.replaceAll(NON_THIN, "").length() / 2);
+        return (str.length() - str.replaceAll(NON_THIN, "").length() / 2);
     }
 
     public static String ellipsize(String text, int max) {
@@ -38,4 +38,5 @@ public class Ellipsizer {
 
         return text.substring(0, end) + "...";
     }
+
 }
